+++ packages/assets/scripts/figma-output-system-icons.js
@@ -1,4 +1,3 @@
-/* eslint-disable max-len */
 /**
  * @typedef {import('@figma-export/types').ComponentOutputter} ComponentOutputter
  * @typedef {import('@figma-export/types').ComponentNode} ComponentNode
@@ -145,7 +144,6 @@ export const icons = ${JSON.stringify(icons, null, 2)};
     await writeFile(output, content, 'utf-8');
     return true;
   } catch (error) {
-    // eslint-disable-next-line no-console
     console.error(`Error reading file ${output}:`, error);
     return false;
   }
@@ -170,7 +168,6 @@ const writeSvgFiles = async (
   try {
     await mkdir(output, { recursive: true });
   } catch (error) {
-    // eslint-disable-next-line no-console
     console.error(`Error creating directory ${output}:`, error);
     return false;
   }
@@ -183,7 +180,6 @@ const writeSvgFiles = async (
         await writeFile(filename, svg, 'utf-8');
         return true;
       } catch (error) {
-        // eslint-disable-next-line no-console
         console.error(`Error writing file ${filename}:`, error);
         return false;
       }