+++ packages/tokens/scripts/build.js
@@ -33,7 +33,7 @@ const { author, name, version } = getPackageInformation();
 /**
  * @type {import('style-dictionary/types').LogConfig['verbosity']}
  */
-const verbosity = 'verbose';
+const verbosity = process.env.CI ? 'default' : 'verbose';
 
 const dictionary = new StyleDictionary({
   log: {
@@ -53,7 +53,6 @@ StyleDictionary.registerFileHeader({
 
 const themes = getAvailableThemes();
 
-// eslint-disable-next-line @typescript-eslint/no-misused-promises
 const cssRuns = themes.map(async ({ mode, theme }) => {
   console.log(`Processing theme: ${theme}, mode: ${mode}`);
   const themeInformation = getInformationForTheme(theme, mode);