+++ packages/tokens/scripts/config.js
@@ -44,56 +44,37 @@ export const DARK_2018_THEME = 'sick2018-dark.json';
 
 /**
  * List of static changed to variables for specific themes and modes.
- * @type {Record<string, OutputVariableChangeList>}
+ * @type {{ global: OutputVariableChangeList } & Record<string, OutputVariableChangeList>}
  */
 export const OUTPUT_VARIABLE_CHANGES = {
-  sick2018_dark: {
+  global: {
     // filled button
-    'syn-button-filled-color-text': 'inherit',
-    'syn-button-filled-color-text-active': 'inherit',
-    'syn-button-filled-color-text-hover': 'inherit',
+    'syn-button-filled-color-text': 'var(--syn-color-neutral-0)',
+    'syn-button-filled-color-text-active': 'var(--syn-color-neutral-0)',
+    'syn-button-filled-color-text-hover': 'var(--syn-color-neutral-0)',
 
     // outlined button
-    'syn-button-outline-color-active': 'none',
-    'syn-button-outline-color-hover': 'none',
-    'syn-button-outline-color-text': 'inherit',
+    'syn-button-outline-color-text': 'var(--syn-color-primary-600)',
     'syn-button-outline-color-text-active': 'var(--syn-color-primary-950)',
     'syn-button-outline-color-text-hover': 'var(--syn-color-primary-900)',
 
-    // Text button
+    // text button
     'syn-button-text-color-text': 'var(--syn-button-color)',
     'syn-button-text-color-text-active': 'var(--syn-button-color-active)',
     'syn-button-text-color-text-hover': 'var(--syn-button-color-hover)',
   },
-  sick2018_light: {
-    // filled button
-    'syn-button-filled-color-text': 'inherit',
-    'syn-button-filled-color-text-active': 'inherit',
-    'syn-button-filled-color-text-hover': 'inherit',
-
+  sick2018_dark: {
     // outlined button
     'syn-button-outline-color-active': 'none',
     'syn-button-outline-color-hover': 'none',
-    'syn-button-outline-color-text': 'inherit',
-    'syn-button-outline-color-text-active': 'var(--syn-color-primary-950)',
-    'syn-button-outline-color-text-hover': 'var(--syn-color-primary-900)',
-
-    // Text button
-    'syn-button-text-color-text': 'var(--syn-button-color)',
-    'syn-button-text-color-text-active': 'var(--syn-button-color-active)',
-    'syn-button-text-color-text-hover': 'var(--syn-button-color-hover)',
+  },
+  sick2018_light: {
+    // outlined button
+    'syn-button-outline-color-active': 'none',
+    'syn-button-outline-color-hover': 'none',
   },
   sick2025_dark: {
   },
   sick2025_light: {
-    // filled button
-    'syn-button-filled-color-text': 'inherit',
-    'syn-button-filled-color-text-active': 'inherit',
-    'syn-button-filled-color-text-hover': 'inherit',
-
-    // Text button
-    'syn-button-text-color-text': 'var(--syn-button-color)',
-    'syn-button-text-color-text-active': 'var(--syn-button-color-active)',
-    'syn-button-text-color-text-hover': 'var(--syn-button-color-hover)',
   },
 };